Updated track controller
