+++ app/include/types.hpp
@@ -12,4 +12,7 @@ struct Buttons {
   Bounce save;
 };
 
+enum class Mode { Replace, Overdub };
+enum class Status { Stop, Record, Play, Pause };
+
 #endif
