+++ app/lib/track-controller/include/types.hpp
@@ -5,13 +5,11 @@
 
 // Bounce objects to easily and reliably read the buttons
 struct Buttons {
-  /* Bounce left;
-  Bounce nav;
-  Bounce select;
-  Bounce right; */
   Bounce stop;
   Bounce record;
   Bounce play;
+  Bounce mode;
+  Bounce save;
 };
 
 enum class Mode { Replace, Overdub };
